Moving everything to exafs.py; Deleting exafs_new.py.
